% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussher-package.R
\docType{package}
\name{ussher-package}
\alias{ussher}
\alias{ussher-package}
\title{ussher: Ussher Data Set Drawn from 1658 Chronology}
\description{
The "ussher" data set is drawn from original chronological textual historic events are classified by index, year, epoch (or one of the 7 ancient "Ages of the World", Biblical source book if referenced (rarely), as well as alternate dating mechanisms, such as "Anno Mundi" (age of the world) or Julian Period. Additional file "usshfull" includes variables that may be of further interest to historians, such as Southern Kingdom and Northern Kingdom discrepant dates, and the original amalgamated Dating mechanic used by Ussher in the original text. The raw data can also be called using "usshraw"
}
\author{
\strong{Maintainer}: Daniel Eness \email{dpauleness@gmail.com} (\href{https://orcid.org/0000-0001-9740-4690}{ORCID})

}
\keyword{internal}
